conquerer: Add rudimentary test case generation to pass/fail branch
The function generate_test_case currently generates positive / negative
values that do not take or take the branch, respectively. The branching
behavior is inverted because Selfie compiles "if" to take a branch if
the case is not true.

generate_test_case distinguishes between point expressions and range
expressions, where point expressions branch on a specific value and
range expressions specify a range that matches.

Range expressions are divided in two parts: Everything before the SLTU
instruction is assumed to be boolean operations while everything else
are integral operations. Right now, everything on the boolean part is
dropped except for "1 - (...)" instrutions because this denotes a
negation. To backtrack the integral part of the range expression, an
inverse operation is performed on the inequality expression until we
reach the end of the trace.

Point expressions have not been implemented yet, but can be implemented
easily by using the same inverse operation as for the range expressions.

Right now, both assume that BEQ compares it to literal zero due to the
way "if" are encoded.